ENH: flatten geofileops API + update docs (#51)
