+++ MANIFEST.in
@@ -3,7 +3,7 @@ include geofileops/util/test.gpkg
 include geofileops/util/geofiletypes.csv
 
 # Include the version.txt file
-include version.txt
+include geofileops/version.txt
 
 # Include all tests so they can be run when packaging the conda package
 graft tests
