+++ docs/development.rst
@@ -2,7 +2,7 @@
 Development
 ===========
 
-The source code can be found on the `GeofileOps git repository`_.
+The source code can be found on the |geofileops git repository|.
 
 If you want to do feature requests or file bug reports, that's the place to 
 be as well.
@@ -23,7 +23,10 @@ dependencies::
     conda install python=3.9 geopandas=0.10 psutil
     conda install pylint pytest rope pydata-sphinx-theme sphinx-automodapi
 
-Now, if you fork the `GeofileOps git repositry`_, you should be able to run/debug the code.
+Now, if you fork the |geofileOps git repository|, you should be able to run/debug the code.
 
 .. _miniconda installer : https://conda.io/projects/conda/en/latest/user-guide/install/index.html
-.. _GeofileOps git repository : https://github.com/geofileops/geofileops
+
+.. |geofileOps git repository| raw:: html
+
+   <a href="https://github.com/geofileops/geofileops" target="_blank">geofileops git repository</a>
