+++ docs/old/reference_automodapi.rst_old
@@ -1,0 +1,8 @@
+
+API Reference
+=============
+
+.. automodapi:: geofileops
+  :no-inheritance-diagram:
+  :skip: Path
+
