--- docs/reference_automodapi.rst
@@ -1,10 +1,0 @@
-
-API Reference
-=============
-.. currentmodule:: geofileops
-
-.. automodapi:: geofileops.geofile
-  :no-inheritance-diagram:
-.. automodapi:: geofileops.geofileops
-  :no-inheritance-diagram:
-
