+++ geofileops/version.txt
@@ -1,0 +1,1 @@
+0.3.2a1
