+++ tests/test_geofile.py
@@ -13,7 +13,7 @@ import shapely.geometry as sh_geom
 
 # Add path so the local geofileops packages are found 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
-from geofileops import geofile
+import geofileops as gfo
 from geofileops.util import geoseries_util
 from geofileops.util.geometry_util import GeometryType
 from geofileops.util import io_util
@@ -24,32 +24,32 @@ def test_add_column(tmpdir):
     # Now add area column
     src = test_helper.TestFiles.polygons_parcels_gpkg
     tmppath = Path(tmpdir) / src.name
-    geofile.copy(src, tmppath)
+    gfo.copy(src, tmppath)
 
     # The area column shouldn't be in the test file yet
-    layerinfo = geofile.get_layerinfo(path=tmppath, layer='parcels')
+    layerinfo = gfo.get_layerinfo(path=tmppath, layer='parcels')
     assert 'AREA' not in layerinfo.columns
         
     ### Add area column ###
     #with test_helper.GdalBin(gdal_installation='gdal_default'):
-    geofile.add_column(tmppath, layer='parcels', name='AREA', type='real', 
+    gfo.add_column(tmppath, layer='parcels', name='AREA', type='real', 
             expression='ST_area(geom)')
         
-    layerinfo = geofile.get_layerinfo(path=tmppath, layer='parcels')
+    layerinfo = gfo.get_layerinfo(path=tmppath, layer='parcels')
     assert 'AREA' in layerinfo.columns
     
-    gdf = geofile.read_file(tmppath)
+    gdf = gfo.read_file(tmppath)
     assert round(gdf['AREA'].astype('float')[0], 1) == round(gdf['OPPERVL'].astype('float')[0], 1)
 
     ### Add perimeter column ###
     #with test_helper.GdalBin(gdal_installation='gdal_default'):
-    geofile.add_column(tmppath, layer='parcels', name='PERIMETER', type=geofile.DataType.REAL, 
+    gfo.add_column(tmppath, layer='parcels', name='PERIMETER', type=gfo.DataType.REAL, 
             expression='ST_perimeter(geom)')
         
-    layerinfo = geofile.get_layerinfo(path=tmppath, layer='parcels')
+    layerinfo = gfo.get_layerinfo(path=tmppath, layer='parcels')
     assert 'AREA' in layerinfo.columns
     
-    gdf = geofile.read_file(tmppath)
+    gdf = gfo.read_file(tmppath)
     assert round(gdf['AREA'].astype('float')[0], 1) == round(gdf['OPPERVL'].astype('float')[0], 1)
 
 def test_cmp(tmpdir):
@@ -68,11 +68,11 @@ def test_cmp(tmpdir):
         
         # Copy test file to tmpdir
         dst = Path(tmpdir) / f"polygons_parcels_output{suffix}"
-        geofile.copy(src, dst)
+        gfo.copy(src, dst)
 
         # Now compare source and dst files
-        assert geofile.cmp(src, dst) == True
-        assert geofile.cmp(src2, dst) == False
+        assert gfo.cmp(src, dst) == True
+        assert gfo.cmp(src2, dst) == False
 
 def test_convert(tmpdir):
     # Prepare test data + run tests
@@ -86,11 +86,11 @@ def test_convert(tmpdir):
         
         # Convert
         dst = Path(tmpdir) / f"polygons_parcels_output{suffix}"
-        geofile.convert(src, dst)
+        gfo.convert(src, dst)
 
         # Now compare source and dst file 
-        src_layerinfo = geofile.get_layerinfo(src)
-        dst_layerinfo = geofile.get_layerinfo(dst)
+        src_layerinfo = gfo.get_layerinfo(src)
+        dst_layerinfo = gfo.get_layerinfo(dst)
         assert src_layerinfo.featurecount == dst_layerinfo.featurecount
         assert len(src_layerinfo.columns) == len(dst_layerinfo.columns)
 
@@ -101,32 +101,32 @@ def test_convert_force_output_geometrytype(tmpdir):
     # Convert polygon test file and force to polygon
     src = test_helper.TestFiles.polygons_parcels_gpkg
     dst = Path(tmpdir) / 'polygons_parcels_to_polygon.gpkg'
-    geofile.convert(src, dst, force_output_geometrytype=GeometryType.POLYGON)
+    gfo.convert(src, dst, force_output_geometrytype=GeometryType.POLYGON)
 
     # Convert polygon test file and force to multipolygon
     src = test_helper.TestFiles.polygons_parcels_gpkg
     dst = Path(tmpdir) / 'polygons_parcels_to_multipolygon.gpkg'
-    geofile.convert(src, dst, force_output_geometrytype=GeometryType.MULTIPOLYGON)
+    gfo.convert(src, dst, force_output_geometrytype=GeometryType.MULTIPOLYGON)
 
     # Convert polygon test file and force to linestring
     src = test_helper.TestFiles.polygons_parcels_gpkg
     dst = Path(tmpdir) / 'polygons_parcels_to_linestring.gpkg'
-    geofile.convert(src, dst, force_output_geometrytype=GeometryType.LINESTRING)
+    gfo.convert(src, dst, force_output_geometrytype=GeometryType.LINESTRING)
 
     # Convert polygon test file and force to multilinestring
     src = test_helper.TestFiles.polygons_parcels_gpkg
     dst = Path(tmpdir) / 'polygons_parcels_to_multilinestring.gpkg'
-    geofile.convert(src, dst, force_output_geometrytype=GeometryType.MULTILINESTRING)
+    gfo.convert(src, dst, force_output_geometrytype=GeometryType.MULTILINESTRING)
 
     # Convert polygon test file and force to point
     src = test_helper.TestFiles.polygons_parcels_gpkg
     dst = Path(tmpdir) / 'polygons_parcels_to_point.gpkg'
-    geofile.convert(src, dst, force_output_geometrytype=GeometryType.POINT)
+    gfo.convert(src, dst, force_output_geometrytype=GeometryType.POINT)
 
     # Convert polygon test file and force to multipoint
     src = test_helper.TestFiles.polygons_parcels_gpkg
     dst = Path(tmpdir) / 'polygons_parcels_to_point.gpkg'
-    geofile.convert(src, dst, force_output_geometrytype=GeometryType.MULTIPOINT)
+    gfo.convert(src, dst, force_output_geometrytype=GeometryType.MULTIPOINT)
 
 def test_copy(tmpdir):
     # Prepare test data + run tests
@@ -140,7 +140,7 @@ def test_copy(tmpdir):
         
         # Copy to dest file
         dst = Path(tmpdir) / f"polygons_parcels_output{suffix}"
-        geofile.copy(src, dst)
+        gfo.copy(src, dst)
         assert src.exists() == True
         assert dst.exists() == True
         if suffix == ".shp":
@@ -149,7 +149,7 @@ def test_copy(tmpdir):
         # Copy to dest dir
         dst_dir = Path(tmpdir) / "dest_dir"
         dst_dir.mkdir(parents=True, exist_ok=True)
-        geofile.copy(src, dst_dir)
+        gfo.copy(src, dst_dir)
         dst = dst_dir / src.name
         assert src.exists() == True
         assert dst.exists() == True
@@ -159,33 +159,33 @@ def test_copy(tmpdir):
 def test_driver_enum():
     ### Test ESRIShapefile Driver ###
     # Test getting a driver for a suffix 
-    geofiletype = geofile.GeofileType('.shp')
-    assert geofiletype == geofile.GeofileType.ESRIShapefile
+    geofiletype = gfo.GeofileType('.shp')
+    assert geofiletype == gfo.GeofileType.ESRIShapefile
 
     # Test getting a driver for a Path
-    path = Path("/testje/path_naar_file.sHp") 
-    geofiletype = geofile.GeofileType(path)
-    assert geofiletype == geofile.GeofileType.ESRIShapefile
+    path = Path("/testje/path_naar_gfo.sHp") 
+    geofiletype = gfo.GeofileType(path)
+    assert geofiletype == gfo.GeofileType.ESRIShapefile
 
     ### GPKG Driver ###
     # Test getting a driver for a suffix 
-    geofiletype = geofile.GeofileType('.gpkg')
-    assert geofiletype == geofile.GeofileType.GPKG
+    geofiletype = gfo.GeofileType('.gpkg')
+    assert geofiletype == gfo.GeofileType.GPKG
 
     # Test getting a driver for a Path
-    path = Path("/testje/path_naar_file.gPkG") 
-    geofiletype = geofile.GeofileType(path)
-    assert geofiletype == geofile.GeofileType.GPKG
+    path = Path("/testje/path_naar_gfo.gPkG") 
+    geofiletype = gfo.GeofileType(path)
+    assert geofiletype == gfo.GeofileType.GPKG
 
     ### SQLite Driver ###
     # Test getting a driver for a suffix 
-    geofiletype = geofile.GeofileType('.sqlite')
-    assert geofiletype == geofile.GeofileType.SQLite
+    geofiletype = gfo.GeofileType('.sqlite')
+    assert geofiletype == gfo.GeofileType.SQLite
 
     # Test getting a driver for a Path
-    path = Path("/testje/path_naar_file.sQlItE") 
-    geofiletype = geofile.GeofileType(path)
-    assert geofiletype == geofile.GeofileType.SQLite
+    path = Path("/testje/path_naar_gfo.sQlItE") 
+    geofiletype = gfo.GeofileType(path)
+    assert geofiletype == gfo.GeofileType.SQLite
 
 def test_get_crs(tmpdir):
     # Prepare test data + run tests
@@ -196,7 +196,7 @@ def test_get_crs(tmpdir):
                 path=test_helper.TestFiles.polygons_parcels_gpkg,
                 tmp_dir=tmp_dir,
                 suffix=suffix)
-        crs = geofile.get_crs(src)
+        crs = gfo.get_crs(src)
         assert crs.to_epsg() == 31370
         
 def test_get_default_layer(tmpdir):
@@ -208,25 +208,8 @@ def test_get_default_layer(tmpdir):
                 path=test_helper.TestFiles.polygons_parcels_gpkg,
                 tmp_dir=tmp_dir,
                 suffix=suffix)
-        layer = geofile.get_default_layer(src)
+        layer = gfo.get_default_layer(src)
         assert layer == 'polygons_parcels-2020'
-
-def test_get_driver(tmpdir):
-    # Prepare test data + run tests
-    tmp_dir = Path(tmpdir)
-    for suffix in test_helper.get_test_suffix_list():
-        # If test input file is in wrong format, convert it
-        src = test_helper.prepare_test_file(
-                path=test_helper.TestFiles.polygons_parcels_gpkg,
-                tmp_dir=tmp_dir,
-                suffix=suffix)
-        
-        if suffix == ".gpkg":
-            assert geofile.get_driver(src) == "GPKG"
-        elif suffix == ".shp":
-            assert geofile.get_driver(src) == "ESRI Shapefile"
-        else:
-            raise Exception(f"test not implemented for suffix {suffix}")
 
 def test_get_layerinfo(tmpdir):
     # Prepare test data + run tests
@@ -250,8 +233,8 @@ def basetest_get_layerinfo(
         layer: Optional[str] = None):
 
     ### Tests on layer specified ###
-    layerinfo = geofile.get_layerinfo(src, layer)
-    assert str(layerinfo).startswith("<class 'geofileops.geofile.LayerInfo'>")
+    layerinfo = gfo.get_layerinfo(src, layer)
+    assert str(layerinfo).startswith("<class 'geofileops.file.LayerInfo'>")
     assert layerinfo.featurecount == 46
     if src.suffix == '.shp':
         assert layerinfo.geometrycolumn == 'geometry'
@@ -259,8 +242,8 @@ def basetest_get_layerinfo(
     elif src.suffix == '.gpkg':
         assert layerinfo.geometrycolumn == 'geom'
         assert layerinfo.name == 'parcels'
-    assert layerinfo.geometrytypename == geofile.GeometryType.MULTIPOLYGON.name
-    assert layerinfo.geometrytype == geofile.GeometryType.MULTIPOLYGON
+    assert layerinfo.geometrytypename == gfo.GeometryType.MULTIPOLYGON.name
+    assert layerinfo.geometrytype == gfo.GeometryType.MULTIPOLYGON
     assert len(layerinfo.columns) == 10
     assert layerinfo.total_bounds is not None
     assert layerinfo.crs is not None
@@ -269,7 +252,7 @@ def basetest_get_layerinfo(
     ### Some tests for exception cases ###
     # Layer specified that doesn't exist
     try:
-        layerinfo = geofile.get_layerinfo(src, "not_existing_layer")
+        layerinfo = gfo.get_layerinfo(src, "not_existing_layer")
         exception_raised = False
     except ValueError:
         exception_raised = True
@@ -278,16 +261,16 @@ def basetest_get_layerinfo(
     # Path specified that doesn't exist
     try:
         not_existing_path = io_util.with_stem(src, "not_existing_layer")
-        layerinfo = geofile.get_layerinfo(not_existing_path)
+        layerinfo = gfo.get_layerinfo(not_existing_path)
         exception_raised = False
     except ValueError:
         exception_raised = True
     assert exception_raised is True
 
     # Multiple layers available, but no layer specified
-    if len(geofile.listlayers(src)) > 1:
+    if len(gfo.listlayers(src)) > 1:
         try:
-            layerinfo = geofile.get_layerinfo(src)
+            layerinfo = gfo.get_layerinfo(src)
             exception_raised = False
         except ValueError:
             exception_raised = True
@@ -304,7 +287,7 @@ def test_get_only_layer(tmpdir):
                 tmp_dir=tmp_dir,
                 suffix=suffix)
 
-        layer = geofile.get_only_layer(src)
+        layer = gfo.get_only_layer(src)
         if suffix == ".gpkg":
             assert layer == 'parcels'
         elif suffix == ".shp":
@@ -314,20 +297,20 @@ def test_get_only_layer(tmpdir):
 
     ### Test Geopackage with 2 layers ###
     srcpath = test_helper.TestFiles.polygons_twolayers_gpkg
-    layers = geofile.listlayers(srcpath)
+    layers = gfo.listlayers(srcpath)
     assert len(layers) == 2
     error_raised = False
     try:
-        layer = geofile.get_only_layer(srcpath)
+        layer = gfo.get_only_layer(srcpath)
     except:
         error_raised = True
     assert error_raised is True
 
 def test_is_geofile():
-    assert geofile.is_geofile(test_helper.TestFiles.polygons_parcels_gpkg) == True
-    assert geofile.is_geofile(test_helper.TestFiles.polygons_parcels_gpkg.with_suffix(".shp")) == True
+    assert gfo.is_geofile(test_helper.TestFiles.polygons_parcels_gpkg) == True
+    assert gfo.is_geofile(test_helper.TestFiles.polygons_parcels_gpkg.with_suffix(".shp")) == True
     
-    assert geofile.is_geofile("/test/testje.txt") == False
+    assert gfo.is_geofile("/test/testje.txt") == False
     
 def test_listlayers(tmpdir):
     # Prepare test data + run tests for one layer
@@ -338,7 +321,7 @@ def test_listlayers(tmpdir):
                 path=test_helper.TestFiles.polygons_parcels_gpkg,
                 tmp_dir=tmp_dir,
                 suffix=suffix)
-        layers = geofile.listlayers(src)
+        layers = gfo.listlayers(src)
     
         if suffix == ".gpkg":
             assert layers[0] == "parcels"
@@ -348,7 +331,7 @@ def test_listlayers(tmpdir):
             raise Exception(f"test not implemented for suffix {suffix}")
 
     # Test geopackage 2 layers
-    layers = geofile.listlayers(test_helper.TestFiles.polygons_twolayers_gpkg)
+    layers = gfo.listlayers(test_helper.TestFiles.polygons_twolayers_gpkg)
     assert 'parcels' in layers
     assert 'zones' in layers
 
@@ -362,7 +345,7 @@ def test_move(tmpdir):
                 tmp_dir=tmp_dir,
                 suffix=suffix)
         dst = Path(tmpdir) / f"polygons_parcels_output{suffix}"
-        geofile.move(src, dst)
+        gfo.move(src, dst)
         assert src.exists() == False
         assert dst.exists() == True
         if suffix == ".shp":
@@ -375,7 +358,7 @@ def test_move(tmpdir):
                 suffix=suffix)
         dst_dir = Path(tmpdir) / "dest_dir"
         dst_dir.mkdir(parents=True, exist_ok=True)
-        geofile.move(src, dst_dir)
+        gfo.move(src, dst_dir)
         dst = dst_dir / src.name
         assert src.exists() == False
         assert dst.exists() == True
@@ -387,32 +370,32 @@ def test_update_column(tmpdir):
     # Now add area column
     src = test_helper.TestFiles.polygons_parcels_gpkg
     tmppath = Path(tmpdir) / 'polygons_parcels.gpkg'
-    geofile.copy(src, tmppath)
+    gfo.copy(src, tmppath)
 
     # The area column shouldn't be in the test file yet
-    layerinfo = geofile.get_layerinfo(path=tmppath, layer='parcels')
+    layerinfo = gfo.get_layerinfo(path=tmppath, layer='parcels')
     assert 'area' not in layerinfo.columns
         
     ### Add + update  area column ###
     #with test_helper.GdalBin(gdal_installation='gdal_default'):
-    geofile.add_column(tmppath, layer='parcels', name='AREA', type='real', expression='ST_area(geom)')
-    geofile.update_column(tmppath, name='AreA', expression='ST_area(geom)')
+    gfo.add_column(tmppath, layer='parcels', name='AREA', type='real', expression='ST_area(geom)')
+    gfo.update_column(tmppath, name='AreA', expression='ST_area(geom)')
 
-    layerinfo = geofile.get_layerinfo(path=tmppath, layer='parcels')
+    layerinfo = gfo.get_layerinfo(path=tmppath, layer='parcels')
     assert 'AREA' in layerinfo.columns
-    gdf = geofile.read_file(tmppath)
+    gdf = gfo.read_file(tmppath)
     assert round(gdf['AREA'].astype('float')[0], 1) == round(gdf['OPPERVL'].astype('float')[0], 1)
 
     ### Update column for rows where area > 5 ###
-    geofile.update_column(tmppath, name="AreA", expression="-1", where="area > 4000")
-    gdf = geofile.read_file(tmppath)
+    gfo.update_column(tmppath, name="AreA", expression="-1", where="area > 4000")
+    gdf = gfo.read_file(tmppath)
     gdf_filtered = gdf[gdf["AREA"] == -1]
     assert len(gdf_filtered) == 20
 
     ### Trying to remove column that doesn't exist should raise ValueError ###
     assert "not_existing column" not in layerinfo.columns
     try:
-        geofile.update_column(tmppath, name="not_existing column", expression="ST_area(geom)")
+        gfo.update_column(tmppath, name="not_existing column", expression="ST_area(geom)")
         exception_raised = False
     except ValueError:
         exception_raised = True
@@ -432,28 +415,28 @@ def test_read_file(tmpdir):
 
 def basetest_read_file(srcpath: Path):
     # Test with defaults
-    read_gdf = geofile.read_file(srcpath)
+    read_gdf = gfo.read_file(srcpath)
     assert isinstance(read_gdf, gpd.GeoDataFrame)
     assert len(read_gdf) == 46
 
     # Test no columns
-    read_gdf = geofile.read_file(srcpath, columns=[])
+    read_gdf = gfo.read_file(srcpath, columns=[])
     assert isinstance(read_gdf, gpd.GeoDataFrame)
     assert len(read_gdf) == 46
 
     # Test specific columns (+ test case insensitivity)
     columns = ['OIDN', 'uidn', 'HFDTLT', 'lblhfdtlt', 'GEWASGROEP', 'lengte', 'OPPERVL']
-    read_gdf = geofile.read_file(srcpath, columns=columns)
+    read_gdf = gfo.read_file(srcpath, columns=columns)
     assert len(read_gdf) == 46
     assert len(read_gdf.columns) == (len(columns) + 1)
 
     # Test no geom
-    read_gdf = geofile.read_file_nogeom(srcpath)
+    read_gdf = gfo.read_file_nogeom(srcpath)
     assert isinstance(read_gdf, pd.DataFrame)
     assert len(read_gdf) == 46
 
     # Test ignore_geometry, no columns
-    read_gdf = geofile.read_file_nogeom(srcpath, columns=[])
+    read_gdf = gfo.read_file_nogeom(srcpath, columns=[])
     assert isinstance(read_gdf, pd.DataFrame)
     assert len(read_gdf) == 46
 
@@ -468,13 +451,13 @@ def test_rename_column(tmpdir):
                 suffix=suffix)
         
         # Check if input file is ok
-        orig_layerinfo = geofile.get_layerinfo(test_path)
+        orig_layerinfo = gfo.get_layerinfo(test_path)
         assert "OPPERVL" in orig_layerinfo.columns
         assert "area" not in orig_layerinfo.columns
 
         # Rename
         try:
-            geofile.rename_column(test_path, "OPPERVL", "area")
+            gfo.rename_column(test_path, "OPPERVL", "area")
             exception_raised = False
         except:
             exception_raised = True
@@ -486,7 +469,7 @@ def test_rename_column(tmpdir):
         else:
             # For file types that support rename, check if it worked
             assert exception_raised is False
-            result_layerinfo = geofile.get_layerinfo(test_path)
+            result_layerinfo = gfo.get_layerinfo(test_path)
             assert "OPPERVL" not in result_layerinfo.columns
             assert "area" in result_layerinfo.columns
 
@@ -501,18 +484,18 @@ def test_rename_layer(tmpdir):
                 suffix=suffix)
     
         if suffix == ".gpkg":
-            geofile.rename_layer(src, layer="parcels", new_layer="parcels_renamed")
-            layernames_renamed = geofile.listlayers(path=src)
+            gfo.rename_layer(src, layer="parcels", new_layer="parcels_renamed")
+            layernames_renamed = gfo.listlayers(path=src)
             assert layernames_renamed[0] == "parcels_renamed"
 
         elif suffix == ".shp":
             # Now test rename layer
             try:
-                geofile.rename_layer(
+                gfo.rename_layer(
                         src, 
                         layer="polygons_parcels", 
                         new_layer="polygons_parcels_renamed")
-                layernames_renamed = geofile.listlayers(path=src)
+                layernames_renamed = gfo.listlayers(path=src)
                 assert layernames_renamed[0] == "polygons_parcels_renamed"
             except Exception as ex:
                 assert "rename_layer is not possible" in str(ex) 
@@ -523,14 +506,14 @@ def test_execute_sql(tmpdir):
     # First copy test file to tmpdir
     src = test_helper.TestFiles.polygons_parcels_gpkg
     tmppath = Path(tmpdir) / 'polygons_parcels.gpkg'
-    geofile.copy(src, tmppath)
+    gfo.copy(src, tmppath)
 
     ### Test using execute_sql for creating/dropping indexes ###
     # Create index
-    geofile.execute_sql(path=tmppath, sql_stmt='CREATE INDEX idx_parcels_oidn ON "parcels"("oidn")')
+    gfo.execute_sql(path=tmppath, sql_stmt='CREATE INDEX idx_parcels_oidn ON "parcels"("oidn")')
 
     # Drop index
-    geofile.execute_sql(path=tmppath, sql_stmt='DROP INDEX idx_parcels_oidn')
+    gfo.execute_sql(path=tmppath, sql_stmt='DROP INDEX idx_parcels_oidn')
 
 def test_spatial_index(tmpdir):
     # Prepare test data + run tests for one layer
@@ -543,18 +526,18 @@ def test_spatial_index(tmpdir):
                 suffix=suffix)
     
         # Check if spatial index present:
-        layer = geofile.get_only_layer(src)
-        has_spatial_index = geofile.has_spatial_index(path=src, layer=layer)
+        layer = gfo.get_only_layer(src)
+        has_spatial_index = gfo.has_spatial_index(path=src, layer=layer)
         assert has_spatial_index is True
 
         # Remove spatial index
-        geofile.remove_spatial_index(path=src, layer=layer)
-        has_spatial_index = geofile.has_spatial_index(path=src, layer=layer)
+        gfo.remove_spatial_index(path=src, layer=layer)
+        has_spatial_index = gfo.has_spatial_index(path=src, layer=layer)
         assert has_spatial_index is False
 
         # Create spatial index
-        geofile.create_spatial_index(path=src, layer=layer)
-        has_spatial_index = geofile.has_spatial_index(path=src, layer=layer)
+        gfo.create_spatial_index(path=src, layer=layer)
+        has_spatial_index = gfo.has_spatial_index(path=src, layer=layer)
         assert has_spatial_index is True
 
 def test_to_file(tmpdir):
@@ -571,14 +554,14 @@ def test_to_file(tmpdir):
 
 def basetest_to_file(srcpath, tmppath):
     # Read test file and write to tmppath
-    read_gdf = geofile.read_file(srcpath)
-    geofile.to_file(read_gdf, tmppath)
-    tmp_gdf = geofile.read_file(tmppath)
+    read_gdf = gfo.read_file(srcpath)
+    gfo.to_file(read_gdf, tmppath)
+    tmp_gdf = gfo.read_file(tmppath)
     assert len(read_gdf) == len(tmp_gdf)
 
     # Append the file again to tmppath
-    geofile.to_file(read_gdf, tmppath, append=True)
-    tmp_gdf = geofile.read_file(tmppath)
+    gfo.to_file(read_gdf, tmppath, append=True)
+    tmp_gdf = gfo.read_file(tmppath)
     assert 2*len(read_gdf) == len(tmp_gdf)
 
 def test_to_file_empty(tmpdir):
@@ -602,15 +585,15 @@ def basetest_to_file_empty(
     test_geometrytypes = geoseries_util.get_geometrytypes(test_gdf.geometry)
     assert len(test_geometrytypes) == 1
     output_none_path = output_dir / f"{srcpath.stem}_none{output_suffix}"
-    geofile.to_file(test_gdf, output_none_path)
+    gfo.to_file(test_gdf, output_none_path)
     
     # Now check the result if the data is still the same after being read again
-    test_read_gdf = geofile.read_file(output_none_path)
+    test_read_gdf = gfo.read_file(output_none_path)
     # Result is the same as the original input
     assert test_read_gdf.geometry[0] is None
     assert isinstance(test_read_gdf.geometry[1], sh_geom.Polygon)  
     # The geometrytype of the column in the file is also the same as originaly
-    test_file_geometrytype = geofile.get_layerinfo(output_none_path).geometrytype
+    test_file_geometrytype = gfo.get_layerinfo(output_none_path).geometrytype
     if output_suffix == '.shp':
         assert test_file_geometrytype == GeometryType.MULTIPOLYGON
     else:
@@ -641,15 +624,15 @@ def basetest_to_file_none(
     test_geometrytypes = geoseries_util.get_geometrytypes(test_gdf.geometry)
     assert len(test_geometrytypes) == 1
     output_none_path = output_dir / f"{srcpath.stem}_none{output_suffix}"
-    geofile.to_file(test_gdf, output_none_path)
+    gfo.to_file(test_gdf, output_none_path)
     
     # Now check the result if the data is still the same after being read again
-    test_read_gdf = geofile.read_file(output_none_path)
+    test_read_gdf = gfo.read_file(output_none_path)
     # Result is the same as the original input
     assert test_read_gdf.geometry[0] is None
     assert isinstance(test_read_gdf.geometry[1], sh_geom.Polygon)  
     # The geometrytype of the column in the file is also the same as originaly
-    test_file_geometrytype = geofile.get_layerinfo(output_none_path).geometrytype
+    test_file_geometrytype = gfo.get_layerinfo(output_none_path).geometrytype
     if output_suffix == '.shp':
         assert test_file_geometrytype == GeometryType.MULTIPOLYGON
     else:
@@ -685,10 +668,10 @@ def basetest_to_file_gpd_empty(
             test_gdf.geometry, ignore_empty_geometries=False)
     assert len(test_geometrytypes_includingempty) == 2
     output_empty_path = output_dir / f"{srcpath.stem}_empty{output_suffix}"
-    test_gdf.to_file(output_empty_path, driver=geofile.get_driver_for_ext(output_suffix))
+    test_gdf.to_file(output_empty_path, driver=gfo.GeofileType(output_suffix).ogrdriver)
     
     # Now check the result if the data is still the same after being read again
-    test_read_gdf = geofile.read_file(output_empty_path)
+    test_read_gdf = gfo.read_file(output_empty_path)
     test_read_geometrytypes = geoseries_util.get_geometrytypes(test_read_gdf.geometry)
     assert len(test_gdf) == len(test_read_gdf)
     if output_suffix == '.shp':
@@ -727,15 +710,15 @@ def basetest_to_file_gpd_none(
     test_geometrytypes = geoseries_util.get_geometrytypes(test_gdf.geometry)
     assert len(test_geometrytypes) == 1
     output_none_path = output_dir / f"{input_path.stem}_none{output_suffix}"
-    test_gdf.to_file(output_none_path, driver=geofile.get_driver_for_ext(output_suffix))
+    test_gdf.to_file(output_none_path, driver=gfo.GeofileType(output_suffix).ogrdriver)
     
     # Now check the result if the data is still the same after being read again
-    test_read_gdf = geofile.read_file(output_none_path)
+    test_read_gdf = gfo.read_file(output_none_path)
     # Result is the same as the original input
     assert test_read_gdf.geometry[0] is None
     assert isinstance(test_read_gdf.geometry[1], sh_geom.Polygon)  
     # The geometrytype of the column in the file is also the same as originaly
-    test_file_geometrytype = geofile.get_layerinfo(output_none_path).geometrytype
+    test_file_geometrytype = gfo.get_layerinfo(output_none_path).geometrytype
     if output_suffix == '.shp':
         # Geometrytype of shapefile always returns the multitype  
         assert test_file_geometrytype == test_geometrytypes[0].to_multitype
@@ -758,7 +741,7 @@ def test_remove(tmpdir):
         assert input_path.exists() == True
 
         # Remove and check result
-        geofile.remove(input_path)
+        gfo.remove(input_path)
         assert input_path.exists() == False
 
 if __name__ == '__main__':
@@ -768,7 +751,7 @@ if __name__ == '__main__':
     # Run!
     #test_convert(tmpdir)
     #test_convert_force_output_geometrytype(tmpdir)
-    #test_get_layerinfo(tmpdir)
+    test_get_layerinfo(tmpdir)
     #test_get_only_layer(tmpdir)
     #test_rename_column(tmpdir)
     #test_rename_layer(tmpdir)
@@ -785,4 +768,4 @@ if __name__ == '__main__':
     #test_to_file_empty_shp(tmpdir)
     #test_to_file_none_gpkg(tmpdir)
     #test_to_file_none_shp(tmpdir)
-    test_update_column(tmpdir)
+    #test_update_column(tmpdir)
