+++ tests/test_sqlite_util.py
@@ -9,7 +9,7 @@ import sys
 # Add path so the local geofileops packages are found 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
 from geofileops.util import sqlite_util
-from geofileops.util.geometry_util import GeometryType
+from geofileops import GeometryType
 import test_helper
 
 def test_exec_spatialite_sql(tmpdir):
