--- version.txt
@@ -1,1 +1,0 @@
-0.3.2a1
